+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+### 3.1.0 - 2017-09-24
+ * Add Friend Code commands
+ * Add `config.js.example`
+ * Fix `meme.js` throwing an error
+
 # 3.0.0 - 2017-09-06
  * Initial release
+